refactor(StaticFileService): 싱글톤 패턴을 통해 동시성 문제 제어 + requestRouter에서 request 처리 로직 삭제
requestRouter는 올바른 서비스 객체를 호출해서 일을 시키는 일만 담당하게 될 예정입니다.
