+++ src/main/java/webserver/RequestRouter.java
@@ -3,15 +3,12 @@ package webserver;
 import com.github.jknack.handlebars.internal.lang3.ArrayUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import service.StaticFileService;
 import webserver.httpMock.CustomHttpRequest;
 import webserver.httpMock.CustomHttpResponse;
 import webserver.httpMock.constants.ContentType;
 import webserver.httpMock.constants.StatusCode;
 
-import java.io.File;
-import java.io.IOException;
-import java.net.URL;
-import java.nio.file.Files;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -19,7 +16,7 @@ public class RequestRouter {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
     private static RequestRouter requestRouter;
     private final Map<String, RequestService> requestMap = new HashMap<>() {{
-        put("[^\\s]+\\.(html|js|css|ico|ttf|woff|svg|eot|woff2|png)$", (req, res) -> getFile(req, res));
+        put("[^\\s]+\\.(html|js|css|ico|ttf|woff|svg|eot|woff2|png)$", StaticFileService.get());
     }};
 
     private RequestRouter(){}
@@ -44,32 +41,7 @@ public class RequestRouter {
         res.addToBody(ArrayUtils.toObject("404 not found".getBytes()));
     }
 
-    public void getFile(CustomHttpRequest req, CustomHttpResponse res){
-        String filePath = null;
-        try {
-            URL resoucePath = getClass().getClassLoader().getResource("./static");
-            String fileType = getFileTypeFromUrl(req.getUrl());
-            res.setContentType(ContentType.getContentTypeByFileType(fileType));
 
-            if(req.getUrl().endsWith("ico") || req.getUrl().endsWith("html")){
-                resoucePath = getClass().getClassLoader().getResource("./templates");
-            }
-
-            filePath = resoucePath.getPath();
-            byte[] file = Files.readAllBytes(new File(filePath + req.getUrl()).toPath());
-            res.addToBody(ArrayUtils.toObject(file));
-            res.setStatusCode(StatusCode.OK);
-
-        } catch (IOException e) {
-            logger.error("file not found at " + filePath + req.getUrl());
-            throw new RuntimeException(e);
-        }
-    }
-
-    private String getFileTypeFromUrl(String url){
-        int index = url.lastIndexOf(".");
-        return url.substring(index+1);
-    }
 
 
 }
